Create programme
This programme asks the user to input a message that they would like encrypted, as well as pick a key (number 1 - 25).
The programme then encrypts the message by moving each character up the alphabet according to the key.
Punctuation and characters outside of the alphabet are ignored and copied through to the new message.
The programme then offers a decryption of a secret message.

This was a fun way for me to practice methods and loops. I had some trouble with the decryption method as it kept showing errors for when I could paste in the 'secret message'. I think I have sorted out the main issues and should hopefully work for any message input.